enable more linters from golanci-lint (#333)
- enables more linters from golang ci lint
- fixes some low hanging fruits

+++ testutil/datagen/btcstaking.go
@@ -154,7 +154,7 @@ func GenRandomBTCDelegation(
 		BtcPk:            delBTCPK,
 		Pop:              pop,
 		FpBtcPkList:      fpBTCPKs,
-		StakingTime:      uint32(stakingTime),
+		StakingTime:      stakingTime,
 		StartHeight:      startHeight,
 		EndHeight:        endHeight,
 		TotalSat:         totalSat,
